Added check for existence of secondary nav menu
If not using a secondary nav menu, you end up with an empty <nav> tag which
messes up the layout of the search bar which get pushed below the title/logo.
This also causes the header to be larger than desired and looks terrible.
This patch simply checks for the existence of a secondary menu before it outputs
the <nav> tags, thus not echoing an empty <nav> tag if you don't use a secondary
nav menu.