Remove styling from No Products filter results. (#2027)
* Remove styling from No Products filter results.

Omits the woocommerce-info styles on the filter results notification
when no products match the filter.

* Fix linting errors.

While the linting errors are not related to this PR, they are failing
in the CI. This should fix the issues with the CSS and JS errors.

* Broaden selector to capture all store archives.

Instead of only targeting the WooCommerce store catalog, we also need to
target (omit) the info box on the archive pages as well, since the
archives all display lists of products under the woocommerce setions.

* Update selector to target new class in WC Core.

Added a new class to the "no products found" info output for us to
contextually target more reliably here.

See https://github.com/woocommerce/woocommerce/pull/34362.

* Update info message padding.

Zeroes out the left and right padding on the info banner so it aligns
with the heading. Also removes the unnecessary element portion of the
selector as to not be overly specific.